+++ facade/src/main/scala/react/semanticui/elements/icon/Icon.scala
@@ -5,8 +5,8 @@ import japgolly.scalajs.react.component.Js.RawMounted
 import japgolly.scalajs.react.component.Js.UnmountedMapped
 import japgolly.scalajs.react.internal.Effect.Id
 import japgolly.scalajs.react.vdom.VdomNode
-import japgolly.scalajs.react.raw.React
 import react.common.syntax._
+import react.common.style._
 import react.semanticui._
 import react.semanticui.{ raw => suiraw }
 import react.semanticui.sizes._
@@ -52,6 +52,7 @@ object Icon {
     bordered:  js.UndefOr[Boolean]              = js.undefined,
     circular:  js.UndefOr[Boolean]              = js.undefined,
     className: js.UndefOr[String]               = js.undefined,
+    clazz:     js.UndefOr[Css]                  = js.undefined,
     color:     js.UndefOr[SemanticColor]        = js.undefined,
     corner:    js.UndefOr[IconCorner]           = js.undefined,
     disabled:  js.UndefOr[Boolean]              = js.undefined,
@@ -69,7 +70,7 @@ object Icon {
     p.as           = as.toJs
     p.bordered     = bordered
     p.circular     = circular
-    p.className    = className
+    p.className    = (className, clazz).toJs
     p.color        = color.toJs
     p.corner       = corner.toJs
     p.disabled     = disabled
@@ -90,6 +91,7 @@ object Icon {
     bordered:     js.UndefOr[Boolean]               = js.undefined,
     circular:     js.UndefOr[Boolean]               = js.undefined,
     className:    js.UndefOr[String]                = js.undefined,
+    clazz:        js.UndefOr[Css]                   = js.undefined,
     color:        js.UndefOr[suiraw.SemanticCOLORS] = js.undefined,
     corner:       js.UndefOr[Boolean | String]      = js.undefined,
     disabled:     js.UndefOr[Boolean]               = js.undefined,
@@ -104,10 +106,11 @@ object Icon {
     `aria-label`: js.UndefOr[String]                = js.undefined
   ): IconProps = {
     val p = (new js.Object).asInstanceOf[IconProps]
-    p.as           = as
-    p.bordered     = bordered
-    p.circular     = circular
-    p.className    = className
+    p.as        = as
+    p.bordered  = bordered
+    p.circular  = circular
+    p.className = (className, clazz).toJs
+
     p.color        = color
     p.corner       = corner
     p.disabled     = disabled
@@ -130,49 +133,4 @@ object Icon {
     : UnmountedMapped[Id, IconProps, Null, RawMounted[IconProps, Null], IconProps, Null] =
     component(p)(children: _*)
 
-  object Group {
-
-    @js.native
-    @JSImport("semantic-ui-react", "Icon.Group")
-    object RawComponent extends js.Object
-
-    @js.native
-    trait GroupProps extends js.Object {
-
-      @JSBracketAccess
-      def apply(key: String): js.Any = js.native
-
-      @JSBracketAccess
-      def update(key: String, v: js.Any): Unit = js.native
-      var as: js.UndefOr[AsT]                    = js.native
-      var children: js.UndefOr[React.Node]       = js.native
-      var className: js.UndefOr[String]          = js.native
-      var content: js.UndefOr[React.Node]        = js.native
-      var size: js.UndefOr[suiraw.SemanticSIZES] = js.native
-    }
-
-    def props(
-      as:        js.UndefOr[AsC]          = js.undefined,
-      children:  js.UndefOr[VdomNode]     = js.undefined,
-      className: js.UndefOr[String]       = js.undefined,
-      content:   js.UndefOr[VdomNode]     = js.undefined,
-      size:      js.UndefOr[SemanticSize] = js.undefined
-    ): GroupProps = {
-      val p = as.toJsObject[GroupProps]
-      p.as        = as.toJs
-      p.children  = children.toJs
-      p.content   = content.toJs
-      p.className = className
-      p.content   = content.toJs
-      p.size      = size.toJs
-      p
-    }
-
-    private val component =
-      JsComponent[GroupProps, Children.Varargs, Null](RawComponent)
-
-    def apply(p: GroupProps, children: VdomNode*)
-      : UnmountedMapped[Id, GroupProps, Null, RawMounted[GroupProps, Null], GroupProps, Null] =
-      component(p)(children: _*)
-  }
 }