+++ facade/src/main/scala/react/semanticui/modules/progress/Progress.scala
@@ -11,6 +11,7 @@ import japgolly.scalajs.react.component.Js.UnmountedMapped
 import japgolly.scalajs.react.vdom.VdomNode
 import japgolly.scalajs.react.internal.Effect.Id
 import react.common.syntax._
+import react.common.style._
 import react.semanticui.{ raw => suiraw }
 import react.semanticui._
 import react.semanticui.colors._
@@ -107,6 +108,7 @@ object Progress {
     autoSuccess: js.UndefOr[Boolean]           = js.undefined,
     children:    js.UndefOr[VdomNode]          = js.undefined,
     className:   js.UndefOr[String]            = js.undefined,
+    clazz:       js.UndefOr[Css]               = js.undefined,
     color:       js.UndefOr[SemanticColor]     = js.undefined,
     content:     js.UndefOr[VdomNode]          = js.undefined,
     error:       js.UndefOr[Boolean]           = js.undefined,
@@ -128,7 +130,7 @@ object Progress {
     p.attached    = attached
     p.autoSuccess = autoSuccess
     p.children    = children.toJs
-    p.className   = className
+    p.className   = (className, clazz).toJs
     p.color       = color.toJs
     p.content     = content.toJs
     p.error       = error